refactor(serve): move SyncScanExecution to ingest.rs as methods
Move SyncScanExecution struct, construction, and execution logic from
mod.rs into ingest.rs. build_sync_scan_execution becomes
SyncScanExecution::new, execute_scan_execution becomes
SyncScanExecution::execute, and the async wrapper becomes
SyncScanExecution::run_async. This collocates scan execution with the
ingest logic it depends on, reducing mod.rs responsibilities.
